[FE] refactor: onNextClicked refactor
slide bar에서
onNextClicked 로직 분리
